[constants] Add `googleServicesFile` constant on iOS
Added a `googleServiceFile` constant for iOS that contains the contents of the `GoogleService-Info.plist` file. This provides an abstraction that aims to support accessing the `GoogleService-Info.plist` in the managed Expo client. In the Expo client it reads the contents from the manifest (the googleServicesFile that was set in the app.json). In all other cases it reads the contents of the file that was shipped with the app.